BosonStrings

Documentation for BosonStrings.

BosonStrings.Operator

Base.:*

Base.:+

Base.:+

Base.:-

Base.:/

Base.show

BosonStrings.ione

BosonStrings.Operator — Type

Represent an normal ordered boson operator Each key is a

Base.:* — Method

Base.:*(o1::Operator, o2::Operator)

Multiplication of two operators. Only supports single boson operator for now.

Base.:+ — Method

Base.:+(o1::Operator, o2::Operator)
Base.:+(o::Operator, a::Number)
Base.:+(a::Number, o::Operator)

Addition between operators and numbers

Base.:+ — Method

Base.:+(o::Operator, term::Tuple{Number, Vararg{Tuple{Int,Int,Int}} })

Add a term to an operator. The term is of the form c, (site1, n1, m1), (site2, n2, m2) ... where triplets (site, n, m) represent $a_{site}^{\dagger n} a_{site}^m$. The following example adds the term $a_{1}^{\dagger 1} a_{1}^1 a_{2}^{\dagger 2} a_{2}^3 a_{4}^{\dagger 5} a_{4}^5$ to a 4-bosons operator:

o = Operator(4)
o += 1, (1, 1, 1), (2, 2, 3), (4, 5, 5)

julia> o
(1.0 + 0.0im) (†1)(1)·(†2)(3)·(†0)(0)·(†5)(5)

Base.:- — Method

Base.:-(o::Operator)
Base.:-(o1::Operator, o2::Operator)
Base.:-(o::Operator, a::Number)
Base.:-(a::Number, o::Operator)

Subtraction between operators and numbers

Base.:/ — Method

Base.:/(o::Operator, a::Number)

Divide an operator by a number

Base.show — Method

Base.show(io::IO, o::Operator)

Print an operator in a human-readable format. The operator is printed as a sum of terms, each term is a coefficient followed by a string of boson operators.

BosonStrings.ione — Method

return the index of the 1 string